Revert deprecation of main_env_export.py
